+++ public/sw-register.js
@@ -1,5 +1,0 @@
-(function swRegister() {
-  if ('serviceWorker' in navigator) {
-    navigator.serviceWorker.register('/sw.js');
-  }
-}());