+++ src/universalMiddleware/index.js
@@ -10,13 +10,18 @@ import App from '../shared/universal/components/App';
  * An express middleware that is capabable of doing React server side rendering.
  */
 function universalReactAppMiddleware(request: $Request, response: $Response) {
+  if (typeof response.locals.nonce !== 'string') {
+    throw new Error('A "nonce" value has not been attached to the response');
+  }
+  const nonce = response.locals.nonce;
+
   if (process.env.DISABLE_SSR === 'true') {
     if (process.env.NODE_ENV === 'development') {
       console.log('==> Handling react route without SSR');  // eslint-disable-line no-console
     }
     // SSR is disabled so we will just return an empty html page and will
     // rely on the client to initialize and render the react application.
-    const html = render();
+    const html = render({ nonce });
     response.status(200).send(html);
     return;
   }
@@ -35,11 +40,17 @@ function universalReactAppMiddleware(request: $Request, response: $Response) {
     </ServerRouter>
   );
 
+  if (typeof response.locals.nonce !== 'string') {
+    throw new Error('A "nonce" value has not been attached to the response');
+  }
+
   // Render the app to a string.
-  const html = render(
+  const html = render({
     // Provide the full app react element.
-    app
-  );
+    app,
+    // Nonce for allowing inline scripts.
+    nonce,
+  });
 
   // Get the render result from the server render context.
   const renderResult = context.getResult();